Separate type declarations, run_tests(areaLog, tests) for simulation, and run_all_tests() for testcases into separate files
