+++ src/bundles/robot_minigame/index.ts
@@ -10,5 +10,5 @@ export {
   init, create_area, create_rect_area, create_obstacle, create_rect_obstacle, complete_init,
   get_distance, get_color,
   move_forward, move_forward_to_wall, rotate, turn_left, turn_right,
-  should_enter_colors, run_tests
+  should_enter_colors, run_all_tests
 } from './functions';
